refactor(frontend): extract data issues panel logic
DataIssuesPanelContent.vue held 284 script lines covering four unrelated
concerns, none of them testable without mounting the panel. Split into

- data-issues-panel-utils.ts: the filter-value narrowing, the compact
  field subset and the paged request payload, all pure
- use-data-issues-panel-list.ts: the paged list, its append/replace
  paging and the double-fetch guard, plus the derived rows
- use-data-issues-panel-selection.ts: the highlighted-event selection,
  which mirrors the route query rather than holding its own state
- use-data-issues-panel-polling.ts: the remediation poll, the
  sync-completed refetch and the KeepAlive activation gating
- use-panel-filter-engagement.ts: the focus grace period that keeps a
  click on a teleported suggestion from dismissing the drawer

The SFC drops to 101 lines of wiring plus the virtual list, which stays
because it is template plumbing.

No behaviour change.
